📦 removed browserify & underscore
