+++ package.json
@@ -1,41 +1,38 @@
 {
-	"name": "moores-holiday-fireplace",
-	"version": "1.1.0",
-	"description": "Turn rows of looped Moores Cloud Holiday lights into a fire",
-	"private": true,
-	"license": "MIT",
-	"keywords": [
-		"fire",
-		"Moores",
-		"Holiday",
-		"lights",
-		"flicker",
-		"cabin",
-		"LED",
-		"IoT"
-	],
-	"author": {
-		"name": "Mountain/\\Ash",
-		"url": "https://mountainash.id.au/"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/mountainash/moores-holiday-fireplace"
-	},
-	"engines": {
-		"node": ">= 20",
-		"npm": ">= 1.4"
-	},
-	"scripts": {
-		"prestart": "browserify client.js -o public/script.js",
-		"start": "node --env-file .env server.js",
-		"dev": "npm run prestart && node --env-file .env --watch server.js"
-	},
-	"dependencies": {
-		"body-parser": "^1.20.2",
-		"browserify": "^17.0.0",
-		"express": "^4.18.2",
-		"holiday-udp": "^0.1.2",
-		"underscore": "^1.13.6"
-	}
+  "name": "moores-holiday-fireplace",
+  "version": "2.0.0",
+  "description": "Turn rows of looped Moores Cloud Holiday lights into a fire",
+  "private": true,
+  "license": "MIT",
+  "keywords": [
+    "fire",
+    "Moores",
+    "Holiday",
+    "lights",
+    "flicker",
+    "cabin",
+    "LED",
+    "IoT"
+  ],
+  "author": {
+    "name": "Mountain/\\Ash",
+    "url": "https://mountainash.id.au/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mountainash/moores-holiday-fireplace"
+  },
+  "engines": {
+    "node": ">= 20",
+    "npm": ">= 1.4"
+  },
+  "scripts": {
+    "start": "node --env-file .env server.js",
+    "dev": "node --env-file .env --watch server.js"
+  },
+  "dependencies": {
+    "body-parser": "^1.20.2",
+    "express": "^4.18.2",
+    "holiday-udp": "^0.1.2"
+  }
 }
